Move and clean test_interpolate.py
